requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
pymupdf
openai
faiss-cpu
numpy

# Step2+ (upload & parse & rag) will use:
# requests
